Add a lot of functionality to the codebase
The library can now connect and request services from a Device.
A main was also added to handle concurrency needed when interacting
with devices. Code was also refactored and broken out in several
different files to hopefully improve readability.

There is still a lot left to do, but I hope to soon start reading
information from the device!